fix: move difficulties to utils
