+++ src/components/Card/components/CardComponent.tsx
@@ -1,30 +1,8 @@
 import CardButton from './CardButton'
 import './CardComponent.scss'
-export interface Difficulty {
-	color: string
-	name: string
-}
+import { difficulties } from 'utils/difficulties'
 
 const CardComponent = () => {
-	const difficulties: Difficulty[] = [
-		{
-			name: 'easy',
-			color: '#069C56',
-		},
-		{
-			name: 'medium',
-			color: '#FF980E',
-		},
-		{
-			name: 'hard',
-			color: '#FF681E',
-		},
-		{
-			name: 'not_guessed',
-			color: '#D3212C',
-		},
-	]
-
 	return (
 		<div className='card-main'>
 			<div className='show-value'>
