+++ src/models/Difficulty.ts
@@ -1,0 +1,4 @@
+export interface Difficulty {
+	color: string
+	name: string
+}
