+++ src/utils/difficulties.tsx
@@ -1,0 +1,20 @@
+import { Difficulty } from 'models/Difficulty'
+
+export const difficulties: Difficulty[] = [
+	{
+		name: 'easy',
+		color: '#069C56',
+	},
+	{
+		name: 'medium',
+		color: '#FF980E',
+	},
+	{
+		name: 'hard',
+		color: '#FF681E',
+	},
+	{
+		name: 'not_guessed',
+		color: '#D3212C',
+	},
+]
